Move context builder to test utils
